# src/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Collection Node Api Documentation",
    "version": "1.0",
    "title": "Collection Node Api Documentation",
    "termsOfService": "urn:tos",
    "contact": {},
    "license": {
      "name": "Open Souce",
      "url": "https://opensource.org/licenses"
    }
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "paths": {
...
    "/item/{id}": {
      "get": {
        "tags": [
          "item-controller"
        ],
        "summary": "findById",
        "operationId": "findByIdUsingGET",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Item"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "deprecated": false
      },
      "delete": {
        "tags": [
          "item-controller"
        ],
        "summary": "delete",
        "operationId": "deleteUser",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "deprecated": false
      }
    },
...
    "/user/{id}": {
      "get": {
        "tags": [
          "user-controller"
        ],
        "summary": "findById",
        "operationId": "findById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "deprecated": false
      },
      "put": {
        "tags": [
          "user-controller"
        ],
        "summary": "update",
        "operationId": "update",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "User",
            "description": "User",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "deprecated": false
      },
      "delete": {
        "tags": [
          "user-controller"
        ],
        "summary": "remove",
        "operationId": "remove",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "deprecated": false
      }
    }
  },
  "definitions": {
...
    "Item": {
      "title": "Item",
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "active": {
          "type": "boolean",
          "example": true
        },
        "code": {
          "type": "string",
          "example": "abc134"
        },
        "complement": {
          "type": "string",
          "example": 0
        },
        "createDate": {
          "type": "string",
          "format": "date-time",
          "example": "2017-05-12T13:00:05-03:00"
        },
        "currentPrice": {
          "type": "number",
          "example": 33.01
        },
        "description": {
          "type": "string",
          "example": "information of item"
        },
        "evaluation": {
          "type": "number",
          "example": 50
        },
        "flagBox": {
          "type": "boolean",
          "example": true
        },
        "flagCdDvdGdBd": {
          "type": "boolean",
          "example": true
        },
        "flagCrib": {
          "type": "boolean",
          "example": true
        },
        "flagDiscCartridge": {
          "type": "boolean",
          "example": true
        },
        "flagGlover": {
          "type": "boolean",
          "example": true
        },
        "flagInvoice": {
          "type": "boolean",
          "example": false
        },
        "flagManual": {
          "type": "boolean",
          "example": true
        },
        "flagPaintpamphlet": {
          "type": "boolean",
          "example": true
        },
        "flagPoster": {
          "type": "boolean",
          "example": false
        },
        "flagProtetor": {
          "type": "boolean",
          "example": true
        },
        "flagReplica": {
          "type": "boolean",
          "example": false
        },
        "flagSealed": {
          "type": "boolean",
          "example": true
        },
        "genre": {
          "type": "string",
          "example": "FPS"
        },
        "haveGame": {
          "type": "boolean",
          "example": true
        },
        "image": {
          "type": "string",
          "example": "c0554d5ac008ec4e0c34821e878eecc2.jpg"
        },
        "localStorage": {
          "type": "string",
          "example": "table"
        },
        "numberOfPlays": {
          "type": "number",
          "example": 20
        },
        "platform": {
          "type": "string",
          "example": "Nintendo"
        },
        "pricePaid": {
          "type": "number",
          "example": 20.2
        },
        "producer": {
          "type": "string",
          "example": "capcom"
        },
        "progression": {
          "type": "number",
          "example": 100
        },
        "provenance": {
          "type": "string",
          "enum": [
            "Original",
            "Pirate"
          ],
          "example": "Original"
        },
        "publisher": {
          "type": "string",
          "example": "tayto"
        },
        "region": {
          "type": "string",
          "example": "USA"
        },
        "screenshots": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "example": [
            "image_1.jpg",
            "image_2.jpg"
          ]
        },
        "situation": {
          "type": "string",
          "example": "Buy"
        },
        "status": {
          "type": "string",
          "enum": [
            "E",
            "P",
            "C"
          ],
          "example": "C"
        },
        "timeOfPlays": {
          "type": "number",
          "example": 10
        },
        "title": {
          "type": "string",
          "example": "title of item"
        },
        "type": {
          "type": "string",
          "example": "game"
        },
        "user": {
          "type": "number",
          "example": 1
        }
      },
      "required": [
        "title",
        "active",
        "region",
        "type",
        "platform",
        "user"
      ]
    },
    "User": {
      "title": "User",
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": "1"
        },
        "name": {
          "type": "string",
          "example": "michael jackson"
        },
        "date_birth": {
          "type": "string",
          "format": "date-time",
          "example": "2017-05-12T13:00:05-03:00"
        },
        "description": {
          "type": "string",
          "example": "information of user"
        },
        "email": {
          "type": "string",
          "example": "email@eamil.com"
        },
        "photo": {
          "type": "string",
          "example": "image.jpg"
        },
        "password": {
          "type": "string",
          "example": "123456"
        },
        "site": {
          "type": "string",
          "example": "http://www.site.com"
        },
        "phone": {
          "type": "string",
          "example": "8197084623"
        }
      },
      "required": [
        "name",
        "date_birth",
        "description",
        "email",
        "photo",
        "password",
        "site",
        "phone"
      ]
    },
...
  }
}